Split input income variables into separate files
